+++ learning_system_backend/controllers/studentController.js
@@ -1,32 +1,15 @@
-import Student from "../models/student.js"
+import { studentService } from "../services/studentService.js"
 import { ApiError } from "../utils/ApiError.js"
 import { catchAsync } from "../utils/catchAsync.js"
 
 export const getAllStudents = catchAsync(async (req, res) => {
   const { class: className, search } = req.query
-
-  const query = {}
-
-  if (className) {
-    query.class = className
-  }
-
-  if (search) {
-    query.$text = { $search: search }
-  }
-
-  const students = await Student.find(query).select("name class currentStatus lastUpdated").sort("-lastUpdated")
-
+  const students = await studentService.getAllStudents(className, search)
   res.json(students)
 })
 
 export const getStudentById = catchAsync(async (req, res) => {
-  const student = await Student.findById(req.params.id)
-
-  if (!student) {
-    throw new ApiError(404, "Student not found")
-  }
-
+  const student = await studentService.getStudentById(req.params.id)
   res.json(student)
 })
 
@@ -37,18 +20,7 @@ export const addProgressNote = catchAsync(async (req, res) => {
     throw new ApiError(400, "Note and status are required")
   }
 
-  const student = await Student.findById(req.params.id)
-
-  if (!student) {
-    throw new ApiError(404, "Student not found")
-  }
-
-  student.progressHistory.push({ note, status })
-  student.currentStatus = status
-  student.lastUpdated = new Date()
-
-  await student.save()
-
+  const student = await studentService.addProgressNote(req.params.id, note, status)
   res.json(student)
 })
 
@@ -59,11 +31,7 @@ export const createStudent = catchAsync(async (req, res) => {
     throw new ApiError(400, "Name and class are required")
   }
 
-  const student = await Student.create({
-    name,
-    class: className,
-  })
-
+  const student = await studentService.createStudent(name, className)
   res.status(201).json(student)
 })
