+++ learning_system_backend/utils/catchAsync.js
@@ -1,7 +1,6 @@
 export const catchAsync = (fn) => {
-    return (req, res, next) => {
-      fn(req, res, next).catch(next)
-    }
+  return (req, res, next) => {
+    fn(req, res, next).catch(next)
   }
-  
-  
+}
+
